24011 功能新增 [Feat] 1. 新增歌词拖动调节进度 2. UI颜色、布局与动画优化

+++ app/src/main/java/org/ww/dpplayer/player/MusicService.java
@@ -203,7 +203,7 @@ public class MusicService extends Service {
         }
     }
 
-    public void seekTo(int progress) {
+    public void seekTo(long progress) {
         player.seekTo(progress);
     }
